refactor(e2e-support): slice B residual lib splits for #4195
Extract adaptive-polling core, explore-assign sweep/bundled siblings,
and civilian expected-lines part files so residual lib modules stay ≤400
physical lines ahead of the 400/500 CI ratchet.
